+++ Robust.UnitTesting/Server/Maps/MapLoaderTest.cs
@@ -11,6 +11,7 @@ using Robust.Shared.Physics.Dynamics;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.Manager;
 using Robust.Shared.Serialization.Manager.Attributes;
+using Robust.Shared.Utility;
 
 namespace Robust.UnitTesting.Server.Maps
 {
@@ -78,15 +79,10 @@ entities:
         public void TestDataLoadPriority()
         {
             // TODO: Fix after serv3
-            var map = IoCManager.Resolve<IMapManager>();
+            // fix what?
 
             var entMan = IoCManager.Resolve<IEntityManager>();
-
-            var mapId = map.CreateMap();
-            // Yay test bullshit
-            var mapUid = map.GetMapEntityId(mapId);
-            entMan.EnsureComponent<PhysicsMapComponent>(mapUid);
-            entMan.EnsureComponent<BroadphaseComponent>(mapUid);
+            entMan.System<SharedMapSystem>().CreateMap(out var mapId);
 
             var traversal = entMan.System<SharedGridTraversalSystem>();
             traversal.Enabled = false;