+++ Robust.UnitTesting/Server/RobustServerSimulation.cs
@@ -71,8 +71,7 @@ namespace Robust.UnitTesting.Server
         /// <summary>
         /// Adds a new map directly to the map manager.
         /// </summary>
-        EntityUid AddMap(int mapId);
-        EntityUid AddMap(MapId mapId);
+        (EntityUid Uid, MapId MapId) CreateMap();
         EntityUid SpawnEntity(string? protoId, EntityCoordinates coordinates);
         EntityUid SpawnEntity(string? protoId, MapCoordinates coordinates);
     }
@@ -99,18 +98,10 @@ namespace Robust.UnitTesting.Server
             return Collection.Resolve<T>();
         }
 
-        public EntityUid AddMap(int mapId)
+        public (EntityUid Uid, MapId MapId) CreateMap()
         {
-            var mapMan = Collection.Resolve<IMapManager>();
-            mapMan.CreateMap(new MapId(mapId));
-            return mapMan.GetMapEntityId(new MapId(mapId));
-        }
-
-        public EntityUid AddMap(MapId mapId)
-        {
-            var mapMan = Collection.Resolve<IMapManager>();
-            mapMan.CreateMap(mapId);
-            return mapMan.GetMapEntityId(mapId);
+            var uid = Collection.Resolve<IEntityManager>().System<SharedMapSystem>().CreateMap(out var mapId);
+            return (uid, mapId);
         }
 
         public EntityUid SpawnEntity(string? protoId, EntityCoordinates coordinates)