+++ Robust.UnitTesting/Shared/GameObjects/Systems/TransformSystemTests.cs
@@ -17,9 +17,6 @@ namespace Robust.UnitTesting.Shared.GameObjects.Systems
                 .RegisterEntitySystems(f => f.LoadExtraSystemType<AnchoredSystemTests.MoveEventTestSystem>())
                 .InitializeInstance();
 
-            // Adds the map with id 1, and spawns entity 1 as the map entity.
-            sim.AddMap(1);
-
             return sim;
         }
 
@@ -31,7 +28,8 @@ namespace Robust.UnitTesting.Shared.GameObjects.Systems
         {
             var sim = SimulationFactory();
             var entMan = sim.Resolve<IEntityManager>();
-            var ent1 = entMan.SpawnEntity(null, new MapCoordinates(Vector2.Zero, new MapId(1)));
+            var map = sim.CreateMap().MapId;
+            var ent1 = entMan.SpawnEntity(null, new MapCoordinates(Vector2.Zero, map));
 
             entMan.System<AnchoredSystemTests.MoveEventTestSystem>().ResetCounters();
             IoCManager.Resolve<IEntityManager>().GetComponent<TransformComponent>(ent1).LocalPosition = Vector2.One;
@@ -47,7 +45,7 @@ namespace Robust.UnitTesting.Shared.GameObjects.Systems
             var sim = SimulationFactory();
             var entManager = sim.Resolve<IEntityManager>();
             var xformSystem = sim.Resolve<IEntitySystemManager>().GetEntitySystem<SharedTransformSystem>();
-            var mapId = new MapId(1);
+            var mapId = sim.CreateMap().MapId;
 
             var parent = entManager.SpawnEntity(null, new MapCoordinates(Vector2.One, mapId));
             var parentXform = entManager.GetComponent<TransformComponent>(parent);
@@ -83,7 +81,7 @@ namespace Robust.UnitTesting.Shared.GameObjects.Systems
             var sim = SimulationFactory();
             var entManager = sim.Resolve<IEntityManager>();
             var xformSystem = sim.Resolve<IEntitySystemManager>().GetEntitySystem<SharedTransformSystem>();
-            var mapId = new MapId(1);
+            var mapId = sim.CreateMap().MapId;
 
             var parent = entManager.SpawnEntity(null, new MapCoordinates(Vector2.One, mapId));
             var parentXform = entManager.GetComponent<TransformComponent>(parent);